Set p wrap and float; fixes changing font size
The logo on the about page previously looked all scrunched up and differently sized than on the index page. It took a very long time to figure out why. I thought I had the styles set to make sure the font was the same for the logo everywhere.

I debugged this by removing everything from the main body of the about page except the logo. As I was putting everything back, I noticed that as soon as the text passed a certain length, the problem reappeared. It occurred to me that text wrapping was the problem, and it turned out to be so.